Dela upp tripui så att update metoderna fick en egen klass.
